+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Improve Schema kinds
 - Add Schema.prototype.namespaces and attach/detouch methods to add/remove
 - Add types to Schema.prototype.references (in addition to entities)
-- Move checkReferences from Model to Schema
+- Move Model.prototype.checkReferences to Schema.prototype.checkConsistency
 
 ## [1.2.3][] - 2021-05-22